+++ rust/ethers-prometheus/Cargo.toml
@@ -14,6 +14,7 @@ async-trait.workspace = true
 derive-new.workspace = true
 derive_builder.workspace = true
 ethers.workspace = true
+ethers-core.workspace = true
 futures.workspace = true
 log.workspace = true
 maplit.workspace = true
