+++ rust/hyperlane-core/Cargo.toml
@@ -35,7 +35,9 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 sha3 = { workspace = true }
 strum = { workspace = true, optional = true, features = ["derive"] }
+tokio.workspace = true
 thiserror = { workspace = true }
+tracing.workspace = true
 primitive-types = { workspace = true, optional = true }
 solana-sdk = { workspace = true, optional = true }
 tiny-keccak = { workspace = true, features = ["keccak"]}
